remove unexpected blank space. closes #181
The three-column layout is made possible by using two floating blocks
for the sides and a normal fluid block for the center. Some elements
from the center currently align themselves with the bottom of either the
left or the right column, leaving a tall blank space.

To remove this, set all elements in the center block to not clear the
floats.