Add color to the right-most column. closes #1838
The right-most column currently shows white with a brown-ish line up
top. It should be all brown-ish except for the area that shows items.

Add a second background on the right side only when we're using a
3-column layout.

The #sidebar element has a minimum height of 600px. It is too high: most
pages have a smaller right-side area. It causes the right-most column to
display a tall white area uselessly. Change this to a minimum height of
200px. (Let's assume this smaller value is realistic)

Also, while we're modifying 3col.css, vim is freaking out because one
line is missing an LF at the end of the line. Add the missing character
so that the file format is automatically recognized as being CRLF.