Show Canada wordmark according to lang. closes #859
The Canada wordmark image (logo in the top left corner) is displayed the
same in both English and French. Normally on governmental sites, this
image changes between languages to show the current language first (on
the left).

Add the French wordmark image and use it in the French header instead of
the English one.