Change font size to relative. closes #220
The current font sizes set on the body element and on the h[123]
elements of the wiki  are absolute sizes in pixels. This prevents IE
users from changing the font size with the "View->Font Size" feature.

Set the sizes to relative values in 'em' to enable the above-mentioned
feature.