Add missing text for label 'label_log_time'
The plugin "timesheet_plugin" does not add translations for the label
'label_log_time'. However, the site in production seems to already have
text for these labels.

Until the plugin adds the right text, add it to this plugin so that it
displays correctly on the interface. Reuse the text that is currently
shown on the production site.