Add space above issues filter form. closes #707
The filtering form is not visible enough and for some users it looks as
if it was part of the content that is shown above it.

To try and alleviate this, add more space below the form so that it
creates a natural separation between the form and the issues list.